getloc will return location; get date will return current date
